+++ packages/libs/lambda-at-edge/src/routing/redirector.ts
@@ -9,6 +9,8 @@ import {
 import * as http from "http";
 import { CloudFrontRequest } from "aws-lambda";
 import { CloudFrontResultResponse } from "aws-lambda";
+import { addDefaultLocaleToPath } from "./locale-utils";
+import Accept from "@hapi/accept";
 
 /**
  * Whether this is the default trailing slash redirect.
@@ -75,6 +77,8 @@ export function getRedirectPath(
   path: string,
   routesManifest: RoutesManifest
 ): { redirectPath: string; statusCode: number } | null {
+  path = addDefaultLocaleToPath(path, routesManifest);
+
   const redirects: RedirectData[] = routesManifest.redirects;
 
   for (const redirect of redirects) {
@@ -172,4 +176,42 @@ export function getDomainRedirectPath(
     }
   }
   return null;
+}
+
+export function getLanguageRedirect(
+  acceptLanguageHeader: string | undefined,
+  normalisedUri: string,
+  routesManifest: RoutesManifest,
+  manifest: OriginRequestDefaultHandlerManifest
+): string | undefined {
+  // Only redirect root page and if accept language is defined
+  if (normalisedUri !== "/" || !acceptLanguageHeader) {
+    return undefined;
+  }
+
+  if (routesManifest.i18n) {
+    const locales = routesManifest.i18n.locales;
+    const defaultLocale = routesManifest.i18n.defaultLocale;
+    const basePath = routesManifest.basePath;
+
+    const preferredLanguage = Accept.language(acceptLanguageHeader);
+
+    let localeToUse = undefined;
+
+    // Find language in locale that matches preferred language
+    for (const locale of locales) {
+      if (preferredLanguage === locale) {
+        localeToUse = locale;
+        break;
+      }
+    }
+
+    if (!localeToUse || localeToUse === defaultLocale) {
+      return undefined;
+    } else {
+      return `${basePath}/${localeToUse}${manifest.trailingSlash ? "/" : ""}`;
+    }
+  }
+
+  return undefined;
 }