+++ packages/serverless-components/nextjs-component/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.19.0-alpha.29](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/serverless-component@1.19.0-alpha.28...@sls-next/serverless-component@1.19.0-alpha.29) (2021-01-25)
+
+### Features
+
+- **lambda-at-edge:** initial support for routing to locale-specific … ([#884](https://github.com/serverless-nextjs/serverless-next.js/issues/884)) ([b4666d3](https://github.com/serverless-nextjs/serverless-next.js/commit/b4666d331e6dd84e3409d4ef2351ab31afd96953))
+- **lambda-at-edge:** start copying locale-specific page/data files t… ([#883](https://github.com/serverless-nextjs/serverless-next.js/issues/883)) ([8bc889c](https://github.com/serverless-nextjs/serverless-next.js/commit/8bc889cb3242e27057bc5f1260faad4e42ea0cfe))
+- **lambda-at-edge, nextjs-commponent:** rewrite from root using accept-language header, other minor bugfixes for locales ([#885](https://github.com/serverless-nextjs/serverless-next.js/issues/885)) ([828b9dd](https://github.com/serverless-nextjs/serverless-next.js/commit/828b9dde89a9cb2559c58ab4e798ee7cbcbdd5e6))
+
 # [1.19.0-alpha.28](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/serverless-component@1.19.0-alpha.27...@sls-next/serverless-component@1.19.0-alpha.28) (2021-01-19)
 
 **Note:** Version bump only for package @sls-next/serverless-component